[c10][fr] Split cuda and non-cuda fr logic into two cpp file
[ghstack-poisoned]
